// src/pages/shop/interfaces/shop.interface.ts
export interface Shop {
  _id?: string;
  name?: string;
  phoneNo?: string;
  email?: string;
  address?: string;
  websiteName?: string;
  slug?: string;
  domain?: string;
  subDomain?: string;
  category?: string;
  dateString?: string;
  subCategory?: string;
  package?: any;
  theme?: any;
  themeColor?: string;
  owner?: string;
  buildStatus?: string;
  port?: number;
  users?: any[];
}